REFACTOR-#4194: create 'BaseDataframePartition' class
