+++ modin/core/dataframe/pandas/partitioning/partition.py
@@ -13,7 +13,6 @@
 
 """The module defines base interface for a partition of a Modin DataFrame."""
 
-from abc import ABC
 from copy import copy
 
 import pandas
@@ -21,93 +20,15 @@ from pandas.api.types import is_scalar
 
 from modin.pandas.indexing import compute_sliced_len
 from modin.core.storage_formats.pandas.utils import length_fn_pandas, width_fn_pandas
+from modin.core.dataframe.base.partitioning.partition import BaseDataframePartition
 
 
-class PandasDataframePartition(ABC):  # pragma: no cover
+class PandasDataframePartition(BaseDataframePartition):  # pragma: no cover
     """
     An abstract class that is base for any partition class of ``pandas`` storage format.
 
     The class providing an API that has to be overridden by child classes.
     """
-
-    _length_cache = None
-    _width_cache = None
-
-    def get(self):
-        """
-        Get the object wrapped by this partition.
-
-        Returns
-        -------
-        object
-            The object that was wrapped by this partition.
-
-        Notes
-        -----
-        This is the opposite of the classmethod `put`.
-        E.g. if you assign `x = PandasDataframePartition.put(1)`, `x.get()` should
-        always return 1.
-        """
-        pass
-
-    def apply(self, func, *args, **kwargs):
-        """
-        Apply a function to the object wrapped by this partition.
-
-        Parameters
-        ----------
-        func : callable
-            Function to apply.
-        *args : iterable
-            Additional positional arguments to be passed in `func`.
-        **kwargs : dict
-            Additional keyword arguments to be passed in `func`.
-
-        Returns
-        -------
-        PandasDataframePartition
-            New `PandasDataframePartition` object.
-
-        Notes
-        -----
-        It is up to the implementation how `kwargs` are handled. They are
-        an important part of many implementations. As of right now, they
-        are not serialized.
-        """
-        pass
-
-    def add_to_apply_calls(self, func, *args, **kwargs):
-        """
-        Add a function to the call queue.
-
-        Parameters
-        ----------
-        func : callable
-            Function to be added to the call queue.
-        *args : iterable
-            Additional positional arguments to be passed in `func`.
-        **kwargs : dict
-            Additional keyword arguments to be passed in `func`.
-
-        Returns
-        -------
-        PandasDataframePartition
-            New `PandasDataframePartition` object with the function added to the call queue.
-
-        Notes
-        -----
-        This function will be executed when `apply` is called. It will be executed
-        in the order inserted; apply's func operates the last and return.
-        """
-        pass
-
-    def drain_call_queue(self):
-        """Execute all operations stored in the call queue on the object wrapped by this partition."""
-        pass
-
-    def wait(self):
-        """Wait for completion of computations on the object wrapped by the partition."""
-        pass
 
     def to_pandas(self):
         """
@@ -199,47 +120,6 @@ class PandasDataframePartition(ABC):  # pragma: no cover
         return new_obj
 
     @classmethod
-    def put(cls, obj):
-        """
-        Put an object into a store and wrap it with partition object.
-
-        Parameters
-        ----------
-        obj : object
-            An object to be put.
-
-        Returns
-        -------
-        PandasDataframePartition
-            New `PandasDataframePartition` object.
-        """
-        pass
-
-    @classmethod
-    def preprocess_func(cls, func):
-        """
-        Preprocess a function before an `apply` call.
-
-        Parameters
-        ----------
-        func : callable
-            Function to preprocess.
-
-        Returns
-        -------
-        callable
-            An object that can be accepted by `apply`.
-
-        Notes
-        -----
-        This is a classmethod because the definition of how to preprocess
-        should be class-wide. Also, we may want to use this before we
-        deploy a preprocessed function to multiple `PandasDataframePartition`
-        objects.
-        """
-        pass
-
-    @classmethod
     def _length_extraction_fn(cls):
         """
         Return the function that computes the length of the object wrapped by this partition.
@@ -262,38 +142,6 @@ class PandasDataframePartition(ABC):  # pragma: no cover
             The function that computes the width of the object wrapped by this partition.
         """
         return width_fn_pandas
-
-    def length(self):
-        """
-        Get the length of the object wrapped by this partition.
-
-        Returns
-        -------
-        int
-            The length of the object.
-        """
-        if self._length_cache is None:
-            cls = type(self)
-            func = cls._length_extraction_fn()
-            preprocessed_func = cls.preprocess_func(func)
-            self._length_cache = self.apply(preprocessed_func)
-        return self._length_cache
-
-    def width(self):
-        """
-        Get the width of the object wrapped by the partition.
-
-        Returns
-        -------
-        int
-            The width of the object.
-        """
-        if self._width_cache is None:
-            cls = type(self)
-            func = cls._width_extraction_fn()
-            preprocessed_func = cls.preprocess_func(func)
-            self._width_cache = self.apply(preprocessed_func)
-        return self._width_cache
 
     @classmethod
     def empty(cls):
