+++ modin/core/execution/python/implementations/pandas_on_python/partitioning/partition.py
@@ -43,11 +43,7 @@ class PandasOnPythonDataframePartition(PandasDataframePartition):
 
     def __init__(self, data, length=None, width=None, call_queue=None):
         self.data = data
-        if call_queue is None:
-            call_queue = []
-        self.call_queue = call_queue
-        self._length_cache = length
-        self._width_cache = width
+        super().__init__(length, width, call_queue)
 
     def get(self):
         """
